feat!: implement tooltip
